apps/api-gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
sse-starlette==2.1.3
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
slowapi==0.1.9
pyyaml==6.0.2
pynvml==11.5.3

# 数据库
sqlalchemy[asyncio]==2.0.35
aiomysql==0.2.0
pymysql==1.1.1
alembic==1.14.1

# 缓存
redis==5.2.1

# 密码哈希
passlib[bcrypt]==1.7.4
